+++ packages/taro-h5/src/index.js
@@ -13,7 +13,7 @@ import {
 import initNativeApi from './native-api'
 import Component from './component'
 
-export {
+export const Taro = {
   initNativeApi,
   Component,
   getEnv,
@@ -26,15 +26,6 @@ export {
   internal_dynamic_recursive
 }
 
-export default {
-  initNativeApi,
-  Component,
-  getEnv,
-  ENV_TYPE,
-  Events,
-  eventCenter,
-  render,
-  internal_safe_get,
-  internal_safe_set,
-  internal_dynamic_recursive
-}
+initNativeApi(Taro)
+
+export default Taro